@@ -1,10 +1,16 @@
 import React from "react";
 import cx from "classnames";
+import { connect } from "react-redux";
+import { toggleTodo } from "../redux/actions";
 
-const Todo = ({ todo }) => (
+const Todo = ({ todo, toggleTodo }) => (
   <li
     className="todo-item"
-    onClick={() => {} /** dispatches action to toggle todo */}
+    onClick={
+      () => {
+        toggleTodo(todo.id);
+      } /** dispatches action to toggle todo */
+    }
   >
     {todo && todo.completed ? "👌" : "👋"}{" "}
     <span
@@ -18,4 +24,4 @@ const Todo = ({ todo }) => (
   </li>
 );
 
-export default Todo;
+export default connect(null, { toggleTodo })(Todo);
