@@ -1,14 +1,30 @@
-export const getTodoList = store =>
+import { VISIBILITY_FILTERS } from "../constants";
+export const getTodoList = (store) =>
   store && store.todos ? store.todos.allIds : [];
 
-export const getTodoById = (store, id) =>
-  store && store.todos && store.todos.byIds
+export const getTodoById = (store, id) => {
+  console.log(store);
+  return store && store.todos && store.todos.byIds
     ? { ...store.todos.byIds[id], id }
     : {};
+};
 
 /**
  * example of a slightly more complex selector
  * select from store combining information from multiple reducers
  */
-export const getTodos = store =>
-  getTodoList(store).map(id => getTodoById(store, id));
+export const getTodos = (store) =>
+  getTodoList(store).map((id) => getTodoById(store, id));
+
+export const getTodosByVisibilityFilter = (store, visibilityFilter) => {
+  const allTodos = getTodos(store);
+  switch (visibilityFilter) {
+    case VISIBILITY_FILTERS.COMPLETED:
+      return allTodos.filter((todo) => todo.completed);
+    case VISIBILITY_FILTERS.INCOMPLETE:
+      return allTodos.filter((todo) => !todo.completed);
+    case VISIBILITY_FILTERS.ALL:
+    default:
+      return allTodos;
+  }
+};
